requirements/requirements-movies_admin.txt
Django==3.2.9
psycopg2==2.9.3
python-dotenv==0.19.2
split_settings==1.0.0
gunicorn==20.1.0
